Smarter cvmfs remount
When using a thin scratch image (image with no layers), then the coresponding cvmfs repository won't be mounted and remounting an unmounted repository results with a failure.
The solution is to do nothing on remount if cvmfs repository wasn't mounted previously. This is the fix for "docker build" from a thin scratch image.